Removed the index table and divided into brands, category and gender table and updated accessing methods

+++ serverMySQL/Shopping-sql/src/main/java/com/example/Shoppingsql/Repository/GenderRepository.java
@@ -1,8 +1,7 @@
 package com.example.Shoppingsql.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.Shoppingsql.Model.Gender;
 
-public interface GenderRepository extends JpaRepository<Gender, Long> {
+public interface GenderRepository extends JpaRepository<Gender , Integer>{
 }
